requirements.txt
feedparser
requests
beautifulsoup4
pymongo
psycopg2-binary
nltk
scikit-learn
pandas
joblib
apache-airflow
